SDK-91 Extract ModuleManager from Setup
